% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_connectivity.R
\name{analyze_connectivity}
\alias{analyze_connectivity}
\title{Analyze ecological connectivity}
\usage{
analyze_connectivity(habitat_sf)
}
\arguments{
\item{habitat_sf}{Objet sf contenant les habitats.}
}
\value{
Tableau des distances entre patches.
}
\description{
Analyse la connectivite entre patches d'habitats.
}
\examples{
\dontrun{
analyze_connectivity(habitat_sf)
}

}
