% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biodiversity_example.R
\docType{data}
\name{biodiversity_example}
\alias{biodiversity_example}
\title{Biodiversity example dataset}
\format{
A data frame with 5 rows and 4 variables:
\describe{
\item{species}{Species name}
\item{abundance}{Observed abundance}
\item{latitude}{Latitude coordinate}
\item{longitude}{Longitude coordinate}
}
}
\source{
Simulated dataset for package examples.
}
\usage{
biodiversity_example
}
\description{
Example dataset included in agrobiodivR.
}
\keyword{datasets}
