% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversity_indices.R
\name{calculate_diversity_indices}
\alias{calculate_diversity_indices}
\title{Calculate diversity indices}
\usage{
calculate_diversity_indices(
  data,
  species_col = "species",
  abundance_col = "abundance"
)
}
\arguments{
\item{data}{Data frame contenant les especes et abondances.}

\item{species_col}{Nom de la colonne especes.}

\item{abundance_col}{Nom de la colonne abondance.}
}
\value{
Tableau contenant les indices.
}
\description{
Calcule les indices de Shannon, Simpson et la richesse specifique.
}
\examples{
biodiversity_data <- data.frame(
  species = c("A", "B", "A", "C"),
  abundance = c(10, 5, 8, 3)
)

calculate_diversity_indices(biodiversity_data)

}
