% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_landscape_metrics.R
\name{calculate_landscape_metrics}
\alias{calculate_landscape_metrics}
\title{Calculate landscape metrics}
\usage{
calculate_landscape_metrics(raster)
}
\arguments{
\item{raster}{Raster d'occupation du sol.}
}
\value{
Tableau des metriques.
}
\description{
Calcule quelques metriques simples du paysage.
}
