% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_landscapes.R
\name{cluster_landscapes}
\alias{cluster_landscapes}
\title{Cluster landscapes}
\usage{
cluster_landscapes(data, k = 3)
}
\arguments{
\item{data}{Data frame des metriques paysageres.}

\item{k}{Nombre de clusters.}
}
\value{
Resultat du clustering.
}
\description{
Classification des paysages par K-means.
}
