% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate model}
\usage{
evaluate_model(observed, predicted)
}
\arguments{
\item{observed}{Valeurs observees.}

\item{predicted}{Valeurs predites.}
}
\value{
Tableau des performances.
}
\description{
Calcule RMSE et R2.
}
