% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_recommendations.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate recommendations}
\usage{
generate_recommendations(diversity_results, connectivity_results)
}
\arguments{
\item{diversity_results}{Resultats des indices de biodiversite.}

\item{connectivity_results}{Resultats de connectivite.}
}
\value{
Data frame contenant les recommandations.
}
\description{
Genere des recommandations a partir des indicateurs calcules.
}
\examples{
div <- data.frame(
  Shannon = 1.5,
  Simpson = 0.7,
  Richness = 5
)

conn <- data.frame(
  Mean_Distance = 600,
  Connectivity_Index = 0.3
)

generate_recommendations(div, conn)

}
