% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(
  diversity_results,
  landscape_results,
  connectivity_results,
  output_file = "agrobiodiv_report.html"
)
}
\arguments{
\item{diversity_results}{Results of biodiversity indices.}

\item{landscape_results}{Results of landscape metrics.}

\item{connectivity_results}{Results of connectivity analysis.}

\item{output_file}{Name of the HTML file to generate.}
}
\value{
Path of the generated HTML file.
}
\description{
Generate a simple HTML report summarizing the results.
}
\examples{
diversity_results <- data.frame(Shannon = 1.5)
landscape_results <- data.frame(Patches = 10)
connectivity_results <- data.frame(Mean_Distance = 300)

temp_report <- tempfile(fileext = ".html")

generate_report(
  diversity_results,
  landscape_results,
  connectivity_results,
  output_file = temp_report
)

}
