% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_ecological_corridors.R
\name{identify_ecological_corridors}
\alias{identify_ecological_corridors}
\title{Identify ecological corridors}
\usage{
identify_ecological_corridors(habitat_sf, threshold_distance = 500)
}
\arguments{
\item{habitat_sf}{Objet sf contenant les patches d'habitats.}

\item{threshold_distance}{Distance maximale entre deux patches pour creer un corridor.}
}
\value{
Objet sf contenant les corridors potentiels.
}
\description{
Identifie des corridors ecologiques potentiels entre habitats proches.
}
\examples{
if(FALSE){
corridors <- identify_ecological_corridors(habitat_sf, threshold_distance = 500)
}

}
