% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_landcover.R
\name{import_landcover}
\alias{import_landcover}
\title{Import land cover raster}
\usage{
import_landcover(file_path)
}
\arguments{
\item{file_path}{Path to the raster file.}
}
\value{
A terra SpatRaster object.
}
\description{
Import a land cover raster file.
}
\examples{
temp_file <- tempfile(fileext = ".tif")

r <- terra::rast(
  nrows = 10,
  ncols = 10,
  xmin = 0,
  xmax = 10,
  ymin = 0,
  ymax = 10
)

terra::values(r) <- sample(1:4, terra::ncell(r), replace = TRUE)
terra::writeRaster(r, temp_file, overwrite = TRUE)

lc <- import_landcover(temp_file)
lc

}
