% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biodiversity_map.R
\name{plot_biodiversity_map}
\alias{plot_biodiversity_map}
\title{Plot biodiversity map}
\usage{
plot_biodiversity_map(biodiversity_sf, index_col = "Shannon")
}
\arguments{
\item{biodiversity_sf}{Objet sf contenant les donnees de biodiversite.}

\item{index_col}{Nom de la colonne contenant l'indice a cartographier.}
}
\value{
Un graphique ggplot.
}
\description{
Produit une carte simple des indices de biodiversite a partir d'un objet sf.
}
\examples{
\dontrun{
plot_biodiversity_map(biodiversity_sf, index_col = "Shannon")
}

}
