% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connectivity_map.R
\name{plot_connectivity_map}
\alias{plot_connectivity_map}
\title{Plot connectivity map}
\usage{
plot_connectivity_map(habitat_sf, corridors_sf)
}
\arguments{
\item{habitat_sf}{Objet sf contenant les habitats.}

\item{corridors_sf}{Objet sf contenant les corridors ecologiques.}
}
\value{
Un graphique ggplot.
}
\description{
Produit une carte des habitats et des corridors ecologiques.
}
\examples{
library(sf)

habitat_sf <- st_sf(
  id = 1:2,
  geometry = st_sfc(
    st_point(c(0, 0)),
    st_point(c(1, 1))
  ),
  crs = 4326
)

corridors_sf <- st_sf(
  id = 1,
  geometry = st_sfc(
    st_linestring(matrix(c(0, 0, 1, 1), ncol = 2, byrow = TRUE))
  ),
  crs = 4326
)

plot_connectivity_map(habitat_sf, corridors_sf)

}
