% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_rf_model.R
\name{train_rf_model}
\alias{train_rf_model}
\title{Train random forest model}
\usage{
train_rf_model(data, target)
}
\arguments{
\item{data}{Jeu de donnees.}

\item{target}{Variable cible.}
}
\value{
Modele Random Forest.
}
\description{
Entraine un modele Random Forest.
}
